+++ src/main/java/com/quasiris/qsf/exception/DebugType.java
@@ -1,0 +1,17 @@
+package com.quasiris.qsf.exception;
+
+public enum DebugType {
+
+    JSON,
+    OBJECT,
+    STRING;
+
+
+    public boolean isJson() {
+        return this == JSON;
+    }
+
+    public boolean isObject() {
+        return this == OBJECT;
+    }
+}